Extract goal form into its own component
